Made git stop tracking keys.js
adding the rule to ignore this file in the previous commit was not
enough for this file to be ignored. Because that file was already
committed and git was tracking it. Here I used
git rm --cached config/keys.js
command to untrack it. By using "--chached" option, I made sure that
only the track record is removed not the file itself.
Next move is to change the API key because anyone who accesses the repo
can see/use that key at the moment.
Note: at this point I have not risked any real user's data as all the
data on the database is placeholders and for development purpose only.