[ALL COMPONENTS] convert imports/exports to ES5 & make a Jest unit test
In order to have better compatibility with Jest testing framework, make
all modules to import and export using ES5 syntax.

ES5 style import and export is problematic on the client side because
the browsers don't support the syntax. To overcome this, use Browserify
package. It generates a bundle.js file which bundles together all the
client side js code which needs to be sent to the client. Therefore, we
write the client-side in a modular way as well but we add a build step
which generates the bundle.js file. The build step is included in
package.json file.

A unit test was generated for MessageBox object which tests its
constructor and the other functions which are added to its prototype.
In order to run the test you need to run "npm run test" which is defined
in package.json file.
Note: At this point of development, there exists 1 test suit which
contains 3 tests and all of them pass.

Included the github repository of Phaser in this repository in order to
access phaser/dist/phaser.js file from there. Hence the relevant script
in dashboard.js is modified.